+++ assets/js/util.js
@@ -1,4 +1,46 @@
+import { createNormalizer } from "@zag-js/types";
+
 const prevAttrsMap = new WeakMap();
+const propMap = {
+  onFocus: "onFocusin",
+  onBlur: "onFocusout",
+  onChange: "onInput",
+  onDoubleClick: "onDblclick",
+  htmlFor: "for",
+  className: "class",
+  defaultValue: "value",
+  defaultChecked: "checked",
+};
+
+const toStyleString = (style) => {
+  let string = "";
+  for (let key in style) {
+    const value = style[key];
+    if (value === null || value === undefined) continue;
+    if (!key.startsWith("--")) key = key.replace(/[A-Z]/g, (match) => `-${match.toLowerCase()}`);
+    string += `${key}:${value};`;
+  }
+  return string;
+};
+
+export const normalizeProps = createNormalizer((props) => {
+  return Object.entries(props).reduce((acc, [key, value]) => {
+    if (value === undefined) return acc;
+
+    if (key in propMap) {
+      key = propMap[key];
+    }
+
+    if (key === "style" && typeof value === "object") {
+      acc.style = toStyleString(value);
+      return acc;
+    }
+
+    acc[key.toLowerCase()] = value;
+
+    return acc;
+  }, {});
+});
 
 export function spreadProps(node, attrs) {
   const oldAttrs = prevAttrsMap.get(node) || {};
@@ -59,4 +101,11 @@ export function spreadProps(node, attrs) {
   return function cleanup() {
     attrKeys.filter(onEvents).forEach(teardown);
   };
+}
+
+export function renderPart(root, name, api) {
+  const camelizedName = name.replace(/-([a-z])/g, (_match, letter) => letter.toUpperCase());
+
+  const part = root.querySelector(`[data-part='${name}']`);
+  if (part) spreadProps(part, api[`${camelizedName}Props`]);
 }